Added code from Starter Kit
The starter kit is publicly available at (https://github.com/edgarjcfn/koding-game-starterkit)
Even though I'm the author, I'd like to point out that this starter kit
was prepared for the Hackathon and posted on the forum
(https://koding.com/Activity/Post/hello-all-for-the-hackathon-i-will-be-working-on-a-project-that-is-very-similar).
in addition to being retweeted byt the Hackathon organizers. I believe
I'm starting from a leveled field with the other competitors.